refactor: add balance.ts service
